Rename save.cpp to save_legacy.cpp
