+++ src/screens/map_screen.cpp
@@ -23,7 +23,6 @@
 #include "gfx/touchscreen.hpp"
 #include "episode/episodeclass.hpp"
 #include "sfx.hpp"
-#include "save.hpp"
 #include "system.hpp"
 #include "exceptions.hpp"
